SGO_Distribution/projeto/backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==3.14.0
Pillow
WeasyPrint
python-decouple
whitenoise
waitress
pyinstaller
